+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-- Switch `IPriceFeedV2` to `IPriceFeedDispatcher` in `BaseToken` and `CollateralManager`.
+- Switch `IPriceFeedV2` to `IPriceFeedDispatcher` in `BaseToken`.
 - Update `ClearingHouseConfig.setTwapInterval` to support 0 interval.
 
 ## [2.5.0] - TBD