+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.6.0] - TBD
 ### Changed
-- Switch `IPriceFeedV2` to `IPriceFeedDispatcher` in `BaseToken` and `CollateralManager`.
+- Switch `IPriceFeedV2` to `IPriceFeedDispatcher` in `BaseToken`.
 - Update `ClearingHouseConfig.setTwapInterval` to support 0 interval.
 
 ## [2.5.0] - TBD